+++ backend/router/userRouter.js
@@ -2,6 +2,7 @@
 const express = require('express');
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
+const { User, addUser, findUserByEmail } = require('');
 
 const router = express.Router();
 const SECRET_KEY = 'your_secret_key'; // 비밀 키
